Refactor codebase by flattening directory structure and consolidating modules (Razor)
- Flattens `src/ast/` to `src/ast.rs`.
- Flattens `src/codegen/` to `src/codegen.rs`.
- Flattens `src/errors/` to `src/errors.rs` by merging submodules.
- Flattens `src/parser/` to `src/parser.rs` by merging `grammar.rs`, `numerals.rs`, and `recursion.rs`.
- Moves `src/parser/grammar.pest` to `src/grammar.pest`.
- Consolidates `src/semantic` by merging `assembly_model.rs` into `model.rs` and inlining test files.
- Ensures all tests pass and code compiles with `cargo check`.
- Updates imports across the codebase to reflect the new structure.

This change adheres to the KISS principle by removing deep nesting for modules with few files and reducing file count.
